chore(suite): routerMiddleware tests
