+++ suite/router/jest.config.js
@@ -4,7 +4,6 @@ module.exports = {
     ...baseConfig,
     roots: [
         '<rootDir>/src',
-        '<rootDir>/tests',
         '<rootDir>/../../suite-common/test-utils/__mocks__',
     ],
 };
